+++ src/wasm-lib/kcl/src/fs/wasm.rs
@@ -1,38 +1,13 @@
 //! Functions for interacting with a file system via wasm.
 
-use std::{
-    pin::Pin,
-    task::{Context, Poll},
-};
-
 use anyhow::Result;
 use wasm_bindgen::prelude::wasm_bindgen;
 
 use crate::{
     errors::{KclError, KclErrorDetails},
     fs::FileSystem,
+    wasm::JsFuture,
 };
-
-struct JsFuture(pub wasm_bindgen_futures::JsFuture);
-
-// Safety: WebAssembly will only ever run in a single-threaded context.
-unsafe impl Send for JsFuture {}
-unsafe impl Sync for JsFuture {}
-
-impl std::future::Future for JsFuture {
-    type Output = Result<wasm_bindgen::JsValue, wasm_bindgen::JsValue>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-        let mut pinned: Pin<&mut wasm_bindgen_futures::JsFuture> = Pin::new(&mut self.get_mut().0);
-        pinned.as_mut().poll(cx)
-    }
-}
-
-impl From<js_sys::Promise> for JsFuture {
-    fn from(promise: js_sys::Promise) -> JsFuture {
-        JsFuture(wasm_bindgen_futures::JsFuture::from(promise))
-    }
-}
 
 #[wasm_bindgen(module = "/../../lang/std/fileSystemManager.ts")]
 extern "C" {
